Replace `use-resize-observer` with own code
It's very little code we have to write. `use-resize-observer` was just
used because of its ability to polyfill the feature. It was added in
2020 when support for ResizeObserver was still sparse. But by now all
important browsers support it for many years already, so there is no
need for a polyfill anymore.

The reason why it is removed now is that it has no React 19 support yet,
which I want to update to in upcoming commits.

This dep was added in a998398c1514f69cd8cb71d2167484374ad52bdd