Replace `react-beforeunload` with own code
It only requires minimal code and gets rid of a dependency. Main problem
with that dependency is that it did not support React 19 yet, which is
the reason for this commit.

The library had support for really old browsers, but since 2023, all
browsers support blocking via `preventDefault()`, so no need to support
other methods still (which are deprecated now anyway).

This helper function will likely be moved to appkit.